refactor: standardize error handling with centralized logger
Create main/logger.js with createLogger() factory.
Replace inconsistent console.warn/error calls across 4 managers
with uniform [module] prefixed log calls.

Also fix broken context-menu.js import paths (from #41).

Closes #50
